Fine tuning implementation of `plot_history`
